move scheduler to its own package named 'sensor'
