+++ pkg/stdoutpublisher/publisher.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/canaryio/canary"
+	"github.com/canaryio/canary/pkg/sensor"
 )
 
 // Publisher implements canary.Publisher, and is our
@@ -17,7 +17,7 @@ func New() *Publisher {
 }
 
 // Publish takes a canary.Measurement and emits data to STDOUT.
-func (p *Publisher) Publish(m canary.Measurement) (err error) {
+func (p *Publisher) Publish(m sensor.Measurement) (err error) {
 	duration := m.Sample.T2.Sub(m.Sample.T1).Seconds() * 1000
 
 	isOK := true
